Add past event recording
Fixed race condition related to the way in which timestamps, hours, and days of the week were retrieved sequetially on separate Calendar instances. Now, all three values are retrieved on the same Calendar. Made Event public, added fluent Event.Builder, and added Event.Instant, which is a 3-tuple of (timestampInMs, hour, day of week). Added Countly#recordPastEvent, which will NOT modify the timestamp/hour/day of week of any Events it is passed. This allows for developers to record events that occured in the past.